Add source maps for ts files and iframe load handlers
