+++ src/app/docfx.ts
@@ -51,4 +51,5 @@ $(() => {
         }
         codeService.init();
 
-})
+});
+
